backend/requirements.txt
# FastAPI e servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Validação
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-cors==1.0.0

# Utilitários
python-dateutil==2.8.2
pytz==2023.3

# Upload de arquivos
aiofiles==23.2.1
pillow==10.2.0

# HTTP requests (para M-Pesa)
httpx==0.26.0
requests==2.31.0

# Logging
loguru==0.7.2
